close to final push, need to kill facebook button later

+++ Roomr/Roomr/Login/UserSetupProfile.swift
@@ -21,8 +21,8 @@ class UserSetupProfile{
     var gender: String
     var genderPref: String
     var housingPref: Int
-    var cleanliness: Float
-    var volume: Float
+    var cleanliness: Int
+    var volume: Int
     var pics : [UIImage]
     var bio: String
     var score: Float
@@ -35,8 +35,8 @@ class UserSetupProfile{
         self.gender = ""
         self.genderPref = ""
         self.housingPref = 0
-        self.cleanliness = 0.0
-        self.volume = 0.0
+        self.cleanliness = 0
+        self.volume = 0
         self.pics = []
         self.bio = ""
         self.score = 0.0
